move topscale stuff outside of the nig crate
