+++ crates/nig/src/params.rs
@@ -16,18 +16,3 @@ pub(crate) struct NigParamsReparametrized {
     pub loc: f64,
     pub log_scale: f64,
 }
-
-#[derive(Debug, Clone, Copy, Serialize)]
-pub struct NigData {
-    pub a: f64,
-    pub b: f64,
-    pub loc: f64,
-    pub scale: f64,
-    pub top_scale: f64,
-}
-
-impl NigData {
-    pub fn params(&self) -> NigParams {
-        NigParams { a: self.a, b: self.b, loc: self.loc, scale: self.scale }
-    }
-}
